Completed Codes .. !!!!

+++ LinqPadDemo/DomainModels/AllEntities.cs
@@ -7,7 +7,7 @@ namespace LinqPadDemo.DomainModels
     {
         public long Id { get; set; }
         public string FullName { get; set; }
-        public byte Age { get; set; }
+        public int Age { get; set; }
         public bool IsActive { get; set; }
         public DateTimeOffset BirthDate { get; set; }
 
@@ -19,7 +19,7 @@ namespace LinqPadDemo.DomainModels
     {
         public long Id { get; set; }
         public string FullName { get; set; }
-        public byte Age { get; set; }
+        public int Age { get; set; }
         public bool IsActive { get; set; }
         public DateTimeOffset BirthDate { get; set; }
